adminController

+++ jblog03/src/main/java/com/poscoict/jblog/vo/CategoryVo.java
@@ -5,7 +5,8 @@ public class CategoryVo {
 	private String name;
 	private String description;
 	private String blogId;
-	private Long count;
+	private int count;
+	
 	public Long getNo() {
 		return no;
 	}
@@ -30,16 +31,17 @@ public class CategoryVo {
 	public void setBlogId(String blogId) {
 		this.blogId = blogId;
 	}
-
-	public Long getCount() {
+	public int getCount() {
 		return count;
 	}
-	public void setCount(Long count) {
+	public void setCount(int count) {
 		this.count = count;
 	}
+
 	@Override
 	public String toString() {
 		return "CategoryVo [no=" + no + ", name=" + name + ", description=" + description + ", blogId=" + blogId + "]";
 	}
+
 	
 }
